+++ server/routes/dev.py
@@ -41,7 +41,6 @@ def screenshot(folder):
 
 @bp.route('/disaster-dashboard')
 def disaster_dashboard():
-    if os.environ.get('FLASK_ENV') != 'autopush' and os.environ.get(
-            'FLASK_ENV') != 'local' and os.environ.get('FLASK_ENV') != 'dev':
+    if not os.environ.get('FLASK_ENV') in ['autopush', 'local', 'dev']:
         flask.abort(404)
     return flask.render_template('dev/disaster_dashboard.html')
